Make event listeners on close-buttons work
Previously, barring the books already displayed when the screen loaded,
each of them having a single event listener attached to them, every time
a person added a book to their library using the 'Add Book' button, the
same event listener would be added to the book being added and to all
the books preceding it on the display. This meant that after adding
several more books, when we tried to remove a book randomly from the
middle of the book list, it would call the event handler function
multiple times because multiple event listeners would be attached to the
close button of that book.

This behavior has since been rectified so that whenever we add a new
book to the library, the event listeners attached to the close buttons
of all the books preceding this book are removed. Then, they are
reattached to the close buttons of these books along with the newly
added book. This process of removing and attaching event listeners
happens every time a new book is added to the library.